% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawum.R
\name{dawum_dataframe}
\alias{dawum_dataframe}
\title{Get DAWUM Data and Return A Data Frame}
\usage{
dawum_dataframe(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Should messages be printed? Default: TRUE}
}
\value{
A data frame with the survey results
}
\description{
This function retrieves data from DAWUM API, extracts relevant data
    from the API response and returns a list object with all metadata and
    survey results.
}
\examples{
\dontrun{dawum_dataframe()}
}
