% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawumr-package.R, R/dawumr.R
\docType{package}
\name{dawumr-package}
\alias{dawumr-package}
\title{dawumr: Retrieve and Prepare Poll Result Data From the DAWUM API}
\description{
This packages provides functions to retrieve and prepare poll result data from the DAWUM API (dawum.de). Data is under the ODC-ODbL license.

This packages provides functions to retrieve and prepare poll result data from the DAWUM API (dawum.de). Data is under the ODC-ODbL license.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{dawumr} package provides convenience functions to pull and transform
election poll data from \href{https://dawum.de/}{dawum.de} using R.

Data from \href{https://dawum.de/}{dawum.de}. Open Database License (ODC-ODbL): \url{https://opendatacommons.org/licenses/odbl/1-0/}
}
\examples{
\dontrun{
# Pull all available poll data as a data frame:
dawum_df <- pull_dawum_dataframe()

# Pull only the latest results by election and institute:
dawum_df <- pull_dawum_dataframe(newest_only = TRUE)
}
}
\keyword{internal}
