% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denormalize.R
\name{denormalize}
\alias{denormalize}
\title{Denormalize the database}
\usage{
denormalize(x, drop_id_vars = TRUE)
}
\arguments{
\item{x}{A list object with the DAWUM database}

\item{drop_id_vars}{Shall the ID variables (e.g. Party ID, Parliament ID) be
dropped after adding the corresponding names? Default: TRUE}
}
\value{
A list object with the DAWUM database, the Result object is denormalized
}
\description{
Denormalize the database
}
