% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_result}
\alias{extract_result}
\title{Wrapper function to extract result data and survey metadata}
\usage{
extract_result(x)
}
\arguments{
\item{x}{A Survey list object from dawum API data}
}
\value{
A dataframe with survey results and metadata
}
\description{
Wrapper function to extract result data and survey metadata
}
