% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_result_dataframe}
\alias{extract_result_dataframe}
\title{Extract survey result data from Survey list object}
\usage{
extract_result_dataframe(x)
}
\arguments{
\item{x}{A Survey list object from dawum API data}
}
\value{
A data frame with the results from a survey
}
\description{
Extract survey result data from Survey list object.
Each record represents the survey result for a party.
}
\examples{
\dontrun{extract_result_dataframe(x)}
}
