% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_result_dataframe}
\alias{generate_result_dataframe}
\title{Generate survey result data from Survey list object}
\usage{
generate_result_dataframe(x)
}
\arguments{
\item{x}{A Survey list object from dawum API data}
}
\value{
A data frame with the results from a survey
}
\description{
Generate survey result data from Survey list object.
    Each record represents the survey result for a party.
}
\examples{
\dontrun{generate_result_dataframe(x)}
}
