% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_result_metadata}
\alias{generate_result_metadata}
\title{Generate metadata from Survey list object}
\usage{
generate_result_metadata(x)
}
\arguments{
\item{x}{A Survey list object from dawum API data}
}
\value{
A data frame with survey metadata (1 row, 8 columns)
}
\description{
Generate metadata from Survey list object
}
\examples{
\dontrun{generate_result_metadata(x)}
}
