% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dawum_data.R
\name{get_dawum_data}
\alias{get_dawum_data}
\title{Get DAWUM Data}
\usage{
get_dawum_data(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Should messages be printed? Default: TRUE}
}
\value{
A list object which contains all metadata and survey results
}
\description{
This function retrieves data from DAWUM API, extracts relevant data
    from the API response and returns a list object with all metadata and
    survey results.
}
\examples{
\dontrun{get_dawum_data()}
}
