% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_tables.R
\name{get_metadata_tables}
\alias{get_metadata_tables}
\title{Get Metadata Tables From API Response}
\usage{
get_metadata_tables(x)
}
\arguments{
\item{x}{A list object containing the fetched API response}
}
\value{
A list object
}
\description{
Get Metadata Tables From API Response
}
\examples{
\dontrun{get_metadata_tables(x)}
}
