% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawum.R
\name{pull_dawum}
\alias{pull_dawum}
\title{Pull DAWUM Data}
\usage{
pull_dawum(newest_only = FALSE, denormalize = FALSE, verbose = TRUE)
}
\arguments{
\item{newest_only}{Logical. Retrieve only the newest data per institute and
parliament/election.}

\item{denormalize}{Logical. Should the names of parties, institutes etc. be
added to the results data frame? Default: TRUE}

\item{verbose}{Logical. Should messages be printed? Default: TRUE}
}
\value{
A list object which contains all metadata and survey results
}
\description{
This function retrieves data from DAWUM API, extracts relevant data
    from the API response and returns a list object with all metadata and
    survey results.
}
\examples{
\dontrun{pull_dawum()}
}
