% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawum.R
\name{pull_dawum_dataframe}
\alias{pull_dawum_dataframe}
\title{Pull DAWUM Data and Return A Data Frame}
\usage{
pull_dawum_dataframe(newest_only = FALSE, verbose = TRUE)
}
\arguments{
\item{newest_only}{Logical. Retrieve only the newest data per institute and
parliament/election.}

\item{verbose}{Logical. Should messages be printed? Default: TRUE}
}
\value{
A data frame with the survey results
}
\description{
This function retrieves data from DAWUM API, extracts relevant data
    from the API response and returns a list object with all metadata and
    survey results.
}
\examples{
\dontrun{dawum_dataframe()}
}
