% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve data}
\usage{
retrieve_data(api_endpoint_url = "https://api.dawum.de/")
}
\arguments{
\item{api_endpoint_url}{The API endpoint url. Default: https://api.dawum.de/}
}
\value{
A list object with the full data provided by the API
}
\description{
Pull data from the API endpoint provided by DAWUM (https://dawum.de/).
Data from dawum.de (License: ODC-ODbL). For the API documentation, refer to
https://dawum.de/API/
}
\examples{
\dontrun{retrieve_data()}
}
