% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_newest_data}
\alias{retrieve_newest_data}
\title{Retrieve newest data only}
\usage{
retrieve_newest_data(
  api_endpoint_url = "https://api.dawum.de/newest_surveys.json"
)
}
\arguments{
\item{api_endpoint_url}{The API endpoint url.
Default: https://api.dawum.de/newest_surveys.json}
}
\value{
A list object with the full data provided by the API
}
\description{
Pull only the newest data per institute and parliament/election
    from the API endpoint provided by DAWUM (https://dawum.de/).
    Data from dawum.de (License: ODC-ODbL). For the API documentation, refer to
    https://dawum.de/API/
}
\examples{
\dontrun{retrieve_newest_data()}
}
